move RecordBatchReader out of geozero code
