+++ src/io/geojson_lines/writer.rs
@@ -3,7 +3,7 @@ use geozero::GeozeroDatasource;
 use std::io::Write;
 
 use crate::error::Result;
-use crate::io::geozero::RecordBatchReader;
+use crate::io::stream::RecordBatchReader;
 
 /// Write a table to newline-delimited GeoJSON
 pub fn write_geojson_lines<W: Write>(table: &mut RecordBatchReader, writer: W) -> Result<()> {
