+++ src/io/geozero/mod.rs
@@ -12,4 +12,3 @@ pub use array::{
     ToMultiPolygonArray, ToPointArray, ToPolygonArray,
 };
 pub use scalar::ToGeometry;
-pub use table::RecordBatchReader;
