+++ src/io/geozero/table/mod.rs
@@ -3,4 +3,3 @@ mod data_source;
 mod json_encoder;
 
 pub use builder::{GeoTableBuilder, GeoTableBuilderOptions};
-pub use data_source::RecordBatchReader;
